Add files via upload
Wanted to use this on my 1366x768 laptop, but didn't want to hardcode it. With a good guess of the centre using outfits and the radius from the ingame HUD, I was getting good results. Had to add a config file to save/retrieve the new settings and keep things simple for users. Setup could use some cleaner instructions, maybe step by step instead and readme needs updated.